+++ projects/ccd-case-ui-toolkit/src/lib/shared/components/palette/judicial-user/write-judicial-user-field.component.ts
@@ -1,10 +1,9 @@
-import { ChangeDetectorRef, Component, OnDestroy, OnInit } from '@angular/core';
+import { Component, OnDestroy, OnInit } from '@angular/core';
 import { FormControl, FormGroup } from '@angular/forms';
-import { ActivatedRoute } from '@angular/router';
 import { Observable, Subscription } from 'rxjs';
 import { catchError, debounceTime, filter, switchMap, tap } from 'rxjs/operators';
 import { JudicialUserModel } from '../../../domain/jurisdiction';
-import { JurisdictionService, SessionStorageService } from '../../../services';
+import { CaseFlagRefdataService, JurisdictionService, SessionStorageService } from '../../../services';
 import { AbstractFieldWriteComponent } from '../base-field';
 
 @Component({
@@ -14,7 +13,6 @@ import { AbstractFieldWriteComponent } from '../base-field';
 })
 export class WriteJudicialUserFieldComponent extends AbstractFieldWriteComponent implements OnInit, OnDestroy {
 
-  private readonly JURISDICTION_ID = 'jid';
   private readonly MINIMUM_SEARCH_CHARACTERS = 2;
   private readonly IDAM_ID = 'idamId';
   private readonly PERSONAL_CODE = 'personalCode';
@@ -22,13 +20,15 @@ export class WriteJudicialUserFieldComponent extends AbstractFieldWriteComponent
   public judicialUserFormGroup: FormGroup;
   public idamIdFormControl: FormControl;
   public personalCodeFormControl: FormControl;
-  public jurisdictionId: string;
+  public jurisdiction: string;
+  public caseType: string;
   public filteredJudicialUsers: JudicialUserModel[] = [];
   public showAutocomplete = false;
   public sub: Subscription;
 
   constructor(private readonly jurisdictionService: JurisdictionService,
-              private readonly sessionStorageService: SessionStorageService) {
+              private readonly sessionStorageService: SessionStorageService,
+              private readonly caseFlagRefDataService: CaseFlagRefdataService) {
     super();
   }
 
@@ -39,12 +39,14 @@ export class WriteJudicialUserFieldComponent extends AbstractFieldWriteComponent
     this.personalCodeFormControl = new FormControl('');
     this.judicialUserFormGroup.addControl(this.PERSONAL_CODE, this.personalCodeFormControl);
 
+    this.setJurisdictionAndCaseType();
+
     this.sub = this.idamIdFormControl.valueChanges.pipe(
       tap(() => this.showAutocomplete = false),
       tap(() => this.filteredJudicialUsers = []),
       debounceTime(300),
       filter((searchTerm: string) => searchTerm && searchTerm.length > this.MINIMUM_SEARCH_CHARACTERS),
-      switchMap((searchTerm: string) => this.filter(searchTerm).pipe(
+      switchMap((searchTerm: string) => this.filterJudicialUsers(searchTerm).pipe(
         tap(() => this.showAutocomplete = true),
         catchError(() => this.filteredJudicialUsers = [])
       ))
@@ -52,13 +54,22 @@ export class WriteJudicialUserFieldComponent extends AbstractFieldWriteComponent
       this.filteredJudicialUsers = judicialUsers;
     });
 
+    console.log('CASE FIELD', this.caseField.value);
+
     if (this.caseField.value?.personalCode) {
       this.loadJudicialUser(this.caseField.value.personalCode);
     }
   }
 
-  public filter(searchTerm: string): Observable<JudicialUserModel[]> {
-    return this.jurisdictionService.searchJudicialUsers(searchTerm, this.getJurisdictionId());
+  public filterJudicialUsers(searchTerm: string): Observable<JudicialUserModel[]> {
+    console.log('SEARCH TERM', searchTerm);
+    return this.caseFlagRefDataService.getHmctsServiceDetailsByCaseType(this.caseType).pipe(
+      // If an error occurs retrieving HMCTS service details by case type ID, try by service name instead
+      catchError(_ => this.caseFlagRefDataService.getHmctsServiceDetailsByServiceName(this.jurisdiction)),
+      // Use switchMap to return an inner Observable of the flag types data, having received the service details
+      // including service_code. This avoids having nested `subscribe`s, which is an anti-pattern!
+      switchMap(serviceDetails => this.jurisdictionService.searchJudicialUsers(searchTerm, serviceDetails[0].service_code))
+    );
   }
 
   public loadJudicialUser(personalCode: string): void {
@@ -71,13 +82,14 @@ export class WriteJudicialUserFieldComponent extends AbstractFieldWriteComponent
     }
   }
 
-  public getJurisdictionId(): string {
+  public setJurisdictionAndCaseType(): void {
     const caseInfoStr = this.sessionStorageService.getItem('caseInfo');
+    console.log('CASE INFO STR', caseInfoStr);
     if (caseInfoStr) {
       const caseInfo = JSON.parse(caseInfoStr);
-      return caseInfo?.jurisdiction;
+      this.jurisdiction = caseInfo?.jurisdiction;
+      this.caseType = caseInfo?.caseType;
     }
-    return null;
   }
 
   public onSelectionChange(judicialUser: JudicialUserModel): void {